Centralized fetchAttributes and other structures in utility file for AddNewStatus and NewItemForm
